chore: refactor superadmin
