+++ api/v1/routes/auth.py
@@ -19,42 +19,6 @@ from api.v1.services.user import user_service
 
 auth = APIRouter(prefix="/auth", tags=["Authentication"])
 
-@auth.post("/login", status_code=status.HTTP_200_OK)
-def login(login_request: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(get_db)):
-    '''Endpoint to log in a user'''
-
-    # Authenticate the user
-    user = user_service.authenticate_user(
-        db=db,
-        username=login_request.username,
-        password=login_request.password
-    )
-
-    # Generate access and refresh tokens
-    access_token = user_service.create_access_token(user_id=user.id)
-    refresh_token = user_service.create_refresh_token(user_id=user.id)
-
-    response = success_response(
-        status_code=200,
-        message='Login successful',
-        data={
-            'access_token': access_token,
-            'token_type': 'bearer',
-        }
-    )
-
-    # Add refresh token to cookies
-    response.set_cookie(
-        key="refresh_token",
-        value=refresh_token,
-        expires=timedelta(days=30),
-        httponly=True,
-        secure=True,
-        samesite="none",
-    )
-
-    return response
-
   
 @auth.post("/register", status_code=status.HTTP_201_CREATED)
 def register(response: Response, user_schema: UserCreate, db: Session = Depends(get_db)):
@@ -85,6 +49,56 @@ def register(response: Response, user_schema: UserCreate, db: Session = Depends(
         key="refresh_token",
         value=refresh_token,
         expires=timedelta(days=60),
+        httponly=True,
+        secure=True,
+        samesite="none",
+    )
+
+    return response
+
+
+@auth.post(path="/register-super-admin", status_code=status.HTTP_201_CREATED)
+def register_as_super_admin(user: UserCreate, db: Session = Depends(get_db)):
+    """Endpoint for super admin creation"""
+
+    user_created = user_service.create_admin(db=db, schema=user)
+    return success_response(
+        status_code=201,
+        message="User Created Successfully",
+        data=user_created.to_dict(),
+    )
+
+
+
+@auth.post("/login", status_code=status.HTTP_200_OK)
+def login(login_request: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(get_db)):
+    '''Endpoint to log in a user'''
+
+    # Authenticate the user
+    user = user_service.authenticate_user(
+        db=db,
+        username=login_request.username,
+        password=login_request.password
+    )
+
+    # Generate access and refresh tokens
+    access_token = user_service.create_access_token(user_id=user.id)
+    refresh_token = user_service.create_refresh_token(user_id=user.id)
+
+    response = success_response(
+        status_code=200,
+        message='Login successful',
+        data={
+            'access_token': access_token,
+            'token_type': 'bearer',
+        }
+    )
+
+    # Add refresh token to cookies
+    response.set_cookie(
+        key="refresh_token",
+        value=refresh_token,
+        expires=timedelta(days=30),
         httponly=True,
         secure=True,
         samesite="none",
